// src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        LoginPeticion: {
            /**
             * Format: email
             * @example ana.perez@villaserena.test
             */
            correo: string;
            /** @example Temporal123 */
            contrasena: string;
        };
// ...
        ReservaAppDetalle: {
            codigo: components["schemas"]["CodigoReserva"];
            entrada: components["schemas"]["FechaEstadia"];
            salida: components["schemas"]["FechaEstadia"];
            estado: components["schemas"]["EstadoReserva"];
            /** @constant */
            horaCheckOut: "12:00";
            tipoHabitacion: components["schemas"]["TipoHabitacionReferencia"];
            numeroHuespedes: number;
            habitacion: components["schemas"]["HabitacionReferencia"] | null;
        };
// ...
    };
// ...
}